[485] Added local java binary check in the Repo

+++ base-vm/build/jenkins/launchLocalBuildEnvironment.ps1
@@ -573,18 +573,23 @@ function Check-Tool
 $jqBinary = (Check-Tool jq.exe $minimalJqVersion $jqInstallerUrl)[1]
 Log_Debug "jqBinary: $jqBinary"
 
-$javaBinary = Get-ChildItem .\java -recurse -include "java.exe"
+# Install Java
+$javaBinary = Get-ChildItem ./java -recurse -include "java.exe"
 Log_Debug "Discovered java binary: `"$javaBinary`""
 
-# Install Java
-Log_Info "Downloading and installing to current directory under ./java"
-$webOutput = "amazon-corretto-windows-x64.zip"
-curl.exe -L --progress-bar $javaInstallerUrl -o .\$webOutput
-Log_Info "Executing... Expand-Archive -LiteralPath .\$webOutput .\java"
-Expand-Archive -LiteralPath .\$webOutput .\java
-$javaBinary = Get-ChildItem .\java -recurse -include "java.exe"
-Log_Debug "Java binary: $javaBinary"
-& $javaBinary -version
+if (  ( Test-Path -Path $javaBinary ) ) {
+    Write-Host "Java Binary already present. Skipping Installation of Java"
+}else{
+    Log_Info "Downloading and installing to current directory under ./java"
+    $webOutput = "amazon-corretto-windows-x64.zip"
+    curl.exe -L --progress-bar $javaInstallerUrl -o .\$webOutput
+    Log_Info "Executing... Expand-Archive -LiteralPath .\$webOutput .\java"
+    Expand-Archive -LiteralPath .\$webOutput .\java
+    $javaBinary = Get-ChildItem .\java -recurse -include "java.exe"
+    Log_Debug "Java binary: $javaBinary"
+    & $javaBinary -version
+}
+
 
 # Create shared workspace directory for Vagrant and Terraform jobs
 $shared_workspace_base_directory_path = $JENKINS_SHARED_WORKSPACE
